+++ src/HttpApi/Controllers/TriggerEventController.php
@@ -4,6 +4,7 @@ namespace BeyondCode\LaravelWebSockets\HttpApi\Controllers;
 
 use BeyondCode\LaravelWebSockets\Dashboard\DashboardLogger;
 use BeyondCode\LaravelWebSockets\Facades\StatisticsLogger;
+use BeyondCode\LaravelWebSockets\WebSockets\WebSocketHook;
 use Illuminate\Http\Request;
 
 class TriggerEventController extends Controller
@@ -22,6 +23,13 @@ class TriggerEventController extends Controller
             ], $request->json()->get('socket_id'));
 
             DashboardLogger::apiMessage(
+                $request->appId,
+                $channelName,
+                $request->json()->get('name'),
+                $request->json()->get('data')
+            );
+
+            app(WebSocketHook::class)->apiMessage(
                 $request->appId,
                 $channelName,
                 $request->json()->get('name'),